Arac guncelleme tamamlandi, code first ile db guncellendi, generic repository'e yeni bir update methodu eklendi, arac entitysi guncellendi, arac form ekranlari duzenlendi.

+++ AracIhale.MODEL/Model/Entities/Arac.cs
@@ -30,7 +30,7 @@ namespace AracIhale.MODEL.Model.Entities
 
         public int ModelID { get; set; }
 
-        public decimal Km { get; set; }
+        public int Km { get; set; }
 
         public int Yil { get; set; }
